requirements.txt
cyhunspell
pandas
fonetika
jellyfish